requirements.txt
fastapi==0.115.11
uvicorn==0.27.1
pydantic>=2.7.4,<3.0.0
langchain==0.3.20
langchain-community==0.3.19
langchain-openai==0.3.8
langgraph==0.2.72
python-dotenv==1.0.1
typing-extensions>=4.8.0 